+++ eniyisinerede.API/ResponseModels/City/CreatedCityResponse.cs
@@ -1,10 +1,7 @@
 ﻿namespace eniyisinerede.API.ResponseModels.City;
 
-public class CreatedCityResponse
+public class CreatedCityResponse : BaseResponseModel
 {
-    public int Id { get; set; }
     public string Name { get; set; }
     public int CountryId { get; set; }
-    public DateTime? CreatedAt { get; set; }
-    public DateTime? UpdatedAt { get; set; }
 }
