+++ eniyisinerede.API/ResponseModels/Country/UpdatedCountryResponse.cs
@@ -1,10 +1,7 @@
 ﻿namespace eniyisinerede.API.ResponseModels.Country;
 
-public class UpdatedCountryResponse
+public class UpdatedCountryResponse : BaseResponseModel
 {
-    public int Id { get; set; }
     public string Name { get; set; }
     public string? Code { get; set; }
-    public DateTime? CreatedAt { get; set; }
-    public DateTime? UpdatedAt { get; set; }
 }
